pyproject.toml
[project]
name = "llm-tavily"
version = "0.1"
description = ""
readme = "README.md"
authors = [{name = "nikkie"}]
license = "Apache-2.0"
classifiers = []
requires-python = ">=3.9"
dependencies = [
    "llm",
    "tavily-python>=0.7.9",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/ftnext/llm-tavily"
Changelog = "https://github.com/ftnext/llm-tavily/releases"
Issues = "https://github.com/ftnext/llm-tavily/issues"
CI = "https://github.com/ftnext/llm-tavily/actions"

[project.entry-points.llm]
tavily = "llm_tavily"

[project.optional-dependencies]
test = ["pytest"]
